refactor: consolidate aws bedrock models, add plugin for nova 2 pro, claude opus 4.6 and remove titan text premier
Refs: RW-1464
