+++ src/Plugin/ocha_ai/Completion/AwsBedrockNova2ProV1.php
@@ -8,19 +8,19 @@ use Drupal\Core\StringTranslation\TranslatableMarkup;
 use Drupal\ocha_ai\Attribute\OchaAiCompletion;
 
 /**
- * AWS Bedrock Nova 2 Lite v1 completion generator.
+ * AWS Bedrock Nova 2 Pro v1 completion generator.
  *
- * Amazon Nova 2 Lite is a fast, cost-effective reasoning model with multimodal
+ * Amazon Nova 2 Pro is a fast, cost-effective reasoning model with multimodal
  * input (text, image, video, documents) and extended context.
  *
  * @see https://docs.aws.amazon.com/bedrock/latest/userguide/model-ids.html
  */
 #[OchaAiCompletion(
-  id: 'aws_bedrock_nova_2_lite_v1',
-  label: new TranslatableMarkup('AWS Bedrock - Nova 2 Lite v1'),
-  description: new TranslatableMarkup('Use AWS Bedrock - Nova 2 Lite v1 as completion generator.')
+  id: 'aws_bedrock_nova_2_pro_v1',
+  label: new TranslatableMarkup('AWS Bedrock - Nova 2 Pro v1'),
+  description: new TranslatableMarkup('Use AWS Bedrock - Nova 2 Pro v1 as completion generator.')
 )]
-class AwsBedrockNova2LiteV1 extends AwsBedrock {
+class AwsBedrockNova2ProV1 extends AwsBedrock {
 
   use Nova2ThinkingTrait;
 
@@ -34,7 +34,7 @@ class AwsBedrockNova2LiteV1 extends AwsBedrock {
    */
   public function defaultConfiguration(): array {
     return [
-      'model' => 'global.amazon.nova-2-lite-v1:0',
+      'model' => 'global.amazon.nova-2-pro-v1:0',
       'max_tokens' => 512,
       'thinking_mode' => 'none',
     ];
@@ -46,9 +46,9 @@ class AwsBedrockNova2LiteV1 extends AwsBedrock {
   public function getModels(): array {
     return [
       // Global endpoint (recommended).
-      'global.amazon.nova-2-lite-v1:0' => $this->t('Amazon - Nova 2 Lite v1 - Global'),
+      'global.amazon.nova-2-pro-v1:0' => $this->t('Amazon - Nova 2 Pro v1 - Global'),
       // Regional endpoint (for data residency / CRIS), 10% more expensive.
-      'us.amazon.nova-2-lite-v1:0' => $this->t('Amazon - Nova 2 Lite v1 - US'),
+      'us.amazon.nova-2-pro-v1:0' => $this->t('Amazon - Nova 2 Pro v1 - US'),
     ];
   }
