+++ src/Plugin/ocha_ai/Completion/AwsBedrockNovaLiteV1.php
@@ -15,7 +15,12 @@ use Drupal\ocha_ai\Attribute\OchaAiCompletion;
   label: new TranslatableMarkup('AWS Bedrock - Nova lite v1'),
   description: new TranslatableMarkup('Use AWS Bedrock - Nova lite v1 as completion generator.')
 )]
-class AwsBedrockNovaLiteV1 extends AwsBedrockNovaV1Base {
+class AwsBedrockNovaLiteV1 extends AwsBedrock {
+
+  /**
+   * {@inheritdoc}
+   */
+  protected bool $supportsFiles = TRUE;
 
   /**
    * {@inheritdoc}
